+++ pandas/core/api.py
@@ -26,7 +26,7 @@ from pandas.core.indexing import IndexSlice
 from pandas.core.dtypes.cast import to_numeric
 from pandas.tseries.offsets import DateOffset
 from pandas.tseries.tools import to_datetime
-from pandas.tseries.index import (DatetimeIndex, Timestamp,
+from pandas.core.indexes.datetimes import (DatetimeIndex, Timestamp,
                                   date_range, bdate_range)
 from pandas.core.indexes.timedeltas import TimedeltaIndex, Timedelta
 from pandas.core.indexes.period import Period, PeriodIndex
