+++ pandas/core/frame.py
@@ -83,7 +83,7 @@ from pandas.util.decorators import Appender, Substitution
 from pandas.util.validators import validate_bool_kwarg
 
 from pandas.core.indexes.period import PeriodIndex
-from pandas.tseries.index import DatetimeIndex
+from pandas.core.indexes.datetimes import DatetimeIndex
 from pandas.core.indexes.timedeltas import TimedeltaIndex
 
 import pandas.core.base as base
