+++ pandas/core/indexes/accessors.py
@@ -12,7 +12,7 @@ from pandas.core.dtypes.common import (
     is_list_like)
 
 from pandas.core.base import PandasDelegate, NoNewAttributesMixin
-from pandas.tseries.index import DatetimeIndex
+from pandas.core.indexes.datetimes import DatetimeIndex
 from pandas._libs.period import IncompatibleFrequency  # noqa
 from pandas.core.indexes.period import PeriodIndex
 from pandas.core.indexes.timedeltas import TimedeltaIndex
