+++ pandas/plotting/_converter.py
@@ -26,7 +26,7 @@ import pandas.core.common as com
 from pandas.core.index import Index
 
 from pandas.core.series import Series
-from pandas.tseries.index import date_range
+from pandas.core.indexes.datetimes import date_range
 import pandas.tseries.tools as tools
 import pandas.tseries.frequencies as frequencies
 from pandas.tseries.frequencies import FreqGroup
