+++ pandas/tests/plotting/test_datetimelike.py
@@ -8,7 +8,7 @@ from pandas.compat import lrange, zip
 import numpy as np
 from pandas import Index, Series, DataFrame
 from pandas.compat import is_platform_mac
-from pandas.tseries.index import date_range, bdate_range
+from pandas.core.indexes.datetimes import date_range, bdate_range
 from pandas.core.indexes.timedeltas import timedelta_range
 from pandas.tseries.offsets import DateOffset
 from pandas.core.indexes.period import period_range, Period, PeriodIndex
