+++ pandas/tests/series/test_analytics.py
@@ -13,7 +13,7 @@ import pandas as pd
 from pandas import (Series, Categorical, DataFrame, isnull, notnull,
                     bdate_range, date_range, _np_version_under1p10)
 from pandas.core.index import MultiIndex
-from pandas.tseries.index import Timestamp
+from pandas.core.indexes.datetimes import Timestamp
 from pandas.core.indexes.timedeltas import Timedelta
 import pandas.core.config as cf
