+++ pandas/tests/series/test_operators.py
@@ -13,7 +13,7 @@ import pandas as pd
 from pandas import (Index, Series, DataFrame, isnull, bdate_range,
                     NaT, date_range, timedelta_range,
                     _np_version_under1p8)
-from pandas.tseries.index import Timestamp
+from pandas.core.indexes.datetimes import Timestamp
 from pandas.core.indexes.timedeltas import Timedelta
 import pandas.core.nanops as nanops
