+++ pandas/tests/tseries/test_offsets.py
@@ -15,7 +15,8 @@ from pandas.core.series import Series
 from pandas.tseries.frequencies import (_offset_map, get_freq_code,
                                         _get_freq_str, _INVALID_FREQ_ERROR,
                                         get_offset, get_standard_freq)
-from pandas.tseries.index import _to_m8, DatetimeIndex, _daterange_cache
+from pandas.core.indexes.datetimes import (
+    _to_m8, DatetimeIndex, _daterange_cache)
 from pandas.tseries.offsets import (BDay, CDay, BQuarterEnd, BMonthEnd,
                                     BusinessHour, WeekOfMonth, CBMonthEnd,
                                     CustomBusinessHour, WeekDay,
