+++ pandas/tseries/api.py
@@ -4,7 +4,7 @@
 
 # flake8: noqa
 
-from pandas.tseries.index import DatetimeIndex, date_range, bdate_range
+from pandas.core.indexes.datetimes import DatetimeIndex, date_range, bdate_range
 from pandas.tseries.frequencies import infer_freq
 from pandas.core.indexes.timedeltas import Timedelta, TimedeltaIndex, timedelta_range
 from pandas.core.indexes.period import Period, PeriodIndex, period_range, pnow
